publish jdcloud-sdk-go version 1.0.4

+++ services/live/models/TrafficStatisticResultData.go
@@ -19,7 +19,7 @@ package models
 
 type TrafficStatisticResultData struct {
 
-    /* 流量，单位：Byte
+    /* 流量，单位:Byte
  (Optional) */
-    Value int `json:"value"`
+    Value int64 `json:"value"`
 }
